feat: 각 섹션 컴포넌트화
