+++ src/pages/orderCheck/components/OrderTrackingSection/OrderTrackingSection.style.ts
@@ -8,5 +8,6 @@ export const section1Container = (theme: Theme) => css`
   height: 100%;
   gap: 1.5rem;
   border: 1px solid ${theme.color.lightgray3};
+  background-color: ${theme.color.white};
   border-radius: 10px;
 `;
