+++ src/pages/orderCheck/components/index.ts
@@ -1,5 +1,15 @@
 import DialButton from "./DialButton/DialButton";
 import PayButton from "./PayButton/PayButton";
 import OrderTrackingSection from "./OrderTrackingSection/OrderTrackingSection";
+import RecentOrderCard from "./RecentOrderCard/RecentOrderCard";
+import OrderNumberSearchSection from "./OrderNumberSearchSection/OrderNumberSearchSection";
+import OrderInfoSection from "./OrderInfoSection/OrderInfoSection";
 
-export { DialButton, PayButton, OrderTrackingSection };
+export {
+  DialButton,
+  PayButton,
+  OrderTrackingSection,
+  RecentOrderCard,
+  OrderNumberSearchSection,
+  OrderInfoSection,
+};
